Fixed bug: noitice message on add new course page Notice: Undefined index: _lp_course_author in ...learnpress\inc\custom-post-types\course.php on line 121

+++ inc/custom-post-types/course.php
@@ -115,16 +115,15 @@ if ( !class_exists( 'LP_Course_Post_Type' ) ) {
 
 		public function update_course( $course_id ) {
 			global $wpdb;
-			$wpdb->update(
-				$wpdb->posts,
-				array( 'post_author' => $_POST['_lp_course_author'] ),
-				array( 'ID' => $course_id )
-			);
-
-			delete_post_meta( $course_id, '_lp_course_author' );
-
+			if( isset( $_POST['_lp_course_author'] ) ) {
+				$wpdb->update(
+					$wpdb->posts,
+					array( 'post_author' => $_POST['_lp_course_author'] ),
+					array( 'ID' => $course_id )
+				);
+				delete_post_meta( $course_id, '_lp_course_author' );
+			}
 		}
-
 
 		function admin_script() {
 			global $post_type;
